+++ src/Hum/Orphans.hs
@@ -2,7 +2,6 @@
 
 module Hum.Orphans where
 
-import qualified Network.MPD as MPD
 import qualified Data.Witherable.Class as W
 import Brick.Widgets.List
 import Control.Lens